support tcp connections (still prefer pipe)

+++ js/controlr.js
@@ -1,4 +1,9 @@
 
+/* global process */
+/* global __dirname */
+
+"use strict"
+
 // switching to bluebird promises (at least temporarily)
 const Promise = require("bluebird");
 
@@ -9,31 +14,36 @@ const path = require( "path" );
 const events = require( "events" );
 const child_process = require( "child_process" );
 
-var start_child_process = function( opts, socket_path ){
+// there's not a lot of well-known stuff operating nearby
+const DEFAULT_TCP_PORT_START = 1440;
+const DEFAULT_TCP_PORT_END = 1480;
+const DEFAULT_TCP_HOST = "127.0.0.1";
+
+var start_child_process = function( opts, args ){
 	
-	var env = {}, file;
-    for( var key in process.env ){ env[key] = process.env[key]; }
-		
+	var env = {}, file, ext = "";
+	for( var key in process.env ){ env[key] = process.env[key]; }
+
 	if( process.platform.match( /win/ )){
 		env.R_HOME = opts.rhome;
 		env.PATH = path.join( opts.rhome, 'bin', 'x64')
 				+ ";" + path.join( opts.basedir ? opts.basedir : __dirname, "..", "build", "Release" )
 				+ ";" + process.env.PATH;
-	 	file = opts.file || path.join( opts.basedir ? opts.basedir : __dirname, "..", "build", "Release", "controlr.exe" );
+		ext = ".exe";
 	}
 	else {
 		env.R_HOME = opts.rhome;
 		env.LD_LIBRARY_PATH = path.join( opts.rhome, 'lib');
-	 	file = opts.file || path.join( opts.basedir ? opts.basedir : __dirname, "..", "build", "Release", "controlr" );
 	}
+
+ 	file = opts.file || path.join( opts.basedir ? opts.basedir : __dirname, "..", "build", "Release", "controlr" + ext );
     
     if( opts.debug ){
         console.info( `file: ${file}` );
         console.info( "env:", env );
     }
     
-    var proc = child_process.spawn( file, 
-        [ socket_path ], { env: env });
+    var proc = child_process.spawn( file, args, { env: env });
     
     if( opts.debug ){
     
@@ -55,31 +65,50 @@ var start_child_process = function( opts, socket_path ){
     
 };
 
+var create_pipe_name = function(){
+
+	var count = 0; // sanity
+	var pipe;
+	do {
+		if( process.platform.match( /win/ )){
+			pipe = "\\\\.\\pipe\\r." + Math.round( 1e8 * Math.random());
+		}
+		else {
+			pipe = path.join( os.tmpdir(), "r." + Math.round( 1e8 * Math.random()));
+		}
+	}
+	while( fs.existsSync(pipe) && count++ < 1024 ); // FIXME: does this actually work on windows with named pipes?  ...
+	
+	if( count >= 1024 ) throw( "Can't create new pipe file" );
+	return pipe;
+}
 
 /**
- * not using this function atm but keep it handy.  returns
- * (via promise) the first available port starting at [port]
- * up to [port+count] or false on no open ports found.
+ * returns (via promise) the first available port starting at [port]
+ * up to [port+count], or rejects.
  */
-var check_port = function( port, count ) {
+var check_port = function( port, host, count ) {
 	return new Promise( function( resolve, reject ){
 		var testserver = net.createServer()
 				.once('error', function (err) {
 					if( count > 0 ){
-						check_port( port + 1, count - 1 ).then( function(x){
-								resolve(x);
+						check_port( port + 1, host, count - 1 ).then( function(x){
+							resolve(x);
+						}).catch(function(e){
+							reject(e);
 						});
 					}
-					else resolve(false);
+					//else resolve(false);
+					else reject( "no ports available in range" );
 				})
 				.once('listening', function() {
 					testserver.once('close', function() { resolve(port); })
 					.close()
 				})
-				.listen(port);
+				.listen(port, host);
 	});
 };
-	 
+	
 var pause = function( delay, func ){
     return new Promise( function( resolve, reject ){
         setTimeout( function(){
@@ -99,14 +128,15 @@ var write_packet = function( socket, packet ){
 
 var on_read = function( socket, buffer, callback ){
     
-    var str = socket.read();
-    if( !str ) return;
+	var str = socket.read();
+	if( !str ) return;
 	 
-    var elts = str.split( "\n" );
-    elts.map( function(x){
-        if( buffer.length ) x = buffer + x;
+	var elts = str.split( "\n" );
+	elts.map( function(x){
+
+		if( buffer.length ) x = buffer + x;
         
-        // valid packet (probably? cheaper than try/catch)
+		// valid packet (probably? cheaper than try/catch)
     	if( x.trim().endsWith( '}' )){		
 			buffer = "";
 			if( x.trim().length ){
@@ -116,12 +146,12 @@ var on_read = function( socket, buffer, callback ){
 					console.info(`parse exception: ${e}`, x);
 					buffer = x;
 				}
-                
 				if( json ) callback(json);
-            }
-        }
-        else buffer = x;
-    });
+			}
+		}
+		else buffer = x;
+	});
+	
 };
     
 var ControlR = function(){
@@ -160,10 +190,14 @@ var ControlR = function(){
             // FIXME: if this is the last call in the application,
             // is there a possibility it will exit before removing 
             // this file?  I think not, but could perhaps verify.
-            fs.unlink( socket_file );
+           	if( socket_file ) fs.unlink( socket_file );
+				  
         });
     };
 
+	 /**
+	  * generic exec function.  
+	  */
     var exec_packet = function( packet ){
         return new Promise( function( resolve, reject ){
             if( busy ) reject( "busy" );
@@ -184,6 +218,9 @@ var ControlR = function(){
             }
         });
     };
+
+	 /** get state */
+	 this.busy = function(){ return busy; }
 
     /** execute a command.  any output will print to the console */
     this.exec = function( cmds ){
@@ -226,58 +263,87 @@ var ControlR = function(){
     this.init = function(){
 
         opts = arguments[0] || {};
-        console.info( "OPTS", opts );
-        
-        return new Promise( function( resolve, reject ){
+        if( opts.debug ) console.info( "controlr init", opts );
+		  
+		  return new Promise( function( resolve, reject ){
             
             if( server ) reject( "already initialized" );
+				var args = [];
+			
+				// default to pipe, but support "tcp".  in the case of tcp, 
+				// additional options [default]:
+				// host ["127.0.0.1"]
+				// port [1440]
+				// last_port [1480]
 				
-            if( process.platform.match( /win/ )){
-	            socket_file = "\\\\.\\pipe\\r." + Math.round( 1e8 * Math.random());
-            }
-            else {
-	            socket_file = path.join( os.tmpdir(), "r." + Math.round( 1e8 * Math.random()));
-            }
+				if( opts.connection_type === "tcp" ){
+					var host = opts.host || DEFAULT_TCP_HOST;
+					var start_port = DEFAULT_TCP_PORT_START;
+					var end_port = DEFAULT_TCP_PORT_END;
+					
+					if( opts.port ){
+						start_port = opts.port;
+						end_port = ( opts.last_port && opts.last_port > opts.port ) ? opts.last_port : opts.port;
+					}
+
+					check_port( start_port, host, end_port - start_port ).then(function(port){
+						args = [host, port];
+   		         server = net.createServer().listen( port, host, connect_callback.bind( this, opts, args, resolve, reject ));
+					}).catch( function(e){
+						reject( e );
+					});
+
+				}
+				else {
+					socket_file = create_pipe_name();
+					args = [socket_file];
+	   	      server = net.createServer().listen( socket_file, connect_callback.bind( this, opts, args, resolve, reject ));
+				}
 				
-            server = net.createServer().listen( socket_file, function(){
-            
-                // set up to catch the child connection
-                server.on( "connection", function(){
-                    if( opts.debug ) console.info( "connection" );
-                    if( socket ) reject( "too many clients" );
-                    
-                    socket = arguments[0];
-                    socket.setEncoding('utf8');
-                    socket.on( "readable", on_read.bind( this, socket, buffer, read_callback ));
-                    socket.on( "close", function(){
-                        if( opts.debug ) console.info( "socket close", arguments );
-                        socket = null;
-                    });
-                    socket.on( "error", function(){
-                        if( opts.debug ) console.info( "socket error", arguments );
-                    });
-
-                    pause( 100 ).then( function(){
-							   if( opts.debug ) console.info( "calling init" );
-                        return exec_packet({
-                            command: 'rinit',
-                            rhome: opts.rhome || "" });
-                    }).then( function(){
-                        if( opts.debug ) console.info( "init complete" ); 
-                        resolve(instance);
-                    }).catch( function(e){
-                        console.info( "Exception in init", e );
-                        reject(e);
-                    });
-                    
-                });
-
-                // then create the child
-                start_child_process( opts, socket_file );
-            });
         });
-    };
-    
+		  
+	 };
+	 
+	 var connect_callback = function(opts, args, resolve, reject){
+            
+		// set up to catch the child connection
+		server.on( "connection", function(){
+						 
+			if( opts.debug ) console.info( "connection" );
+			if( socket ) reject( "too many clients" );
+							
+			socket = arguments[0];
+			socket.setEncoding('utf8');
+			socket.on( "readable", on_read.bind( this, socket, buffer, read_callback ));
+			socket.on( "close", function(){
+				if( opts.debug ) console.info( "socket close", arguments );
+				socket = null;
+			});
+			socket.on( "error", function(){
+				if( opts.debug ) console.info( "socket error", arguments );
+			});
+
+			pause( 100 ).then( function(){
+				if( opts.debug ) console.info( "calling init" );
+				return exec_packet({
+						command: 'rinit',
+						rhome: opts.rhome || "" });
+			}).then( function(){
+				if( opts.debug ) console.info( "init complete" ); 
+				resolve(instance);
+			}).catch( function(e){
+				console.info( "Exception in init", e );
+				reject(e);
+			});
+			
+		});
+
+		// then create the child
+		//start_child_process( opts, [socket_file] );
+		start_child_process( opts, args );
+
+	 };
+	 
     // notwithstanding the above comment, you can init right away if you want
     if( arguments[0] ) this.init.apply( this, arguments );
